// web-application/src/types-openapi/api/models/UpdateUserRequest.ts
export interface UpdateUserRequest {
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    userId: string;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    fullName: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    userName: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    email: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    password?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    phoneNumber: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    role: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    address?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    city?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    district?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    ward?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof UpdateUserRequest
     */
    dateOfBirth?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof UpdateUserRequest
     */
    note?: string | null;
}